# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: MC Roleplay Server API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /forms:
    get:
      operationId: listForms
      tags: [forms]
      summary: List forms and polls (admin sees all, members see open only)
      responses:
        "200":
          description: List of forms/polls
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FormSummary"
    post:
      operationId: createForm
      tags: [forms]
      summary: Admin creates a form or poll
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateFormInput"
      responses:
        "201":
          description: Created form
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FormDetail"
  /forms/{id}:
    get:
      operationId: getForm
      tags: [forms]
      summary: Get form/poll detail including fields or options
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Form detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FormDetail"
    patch:
      operationId: updateForm
      tags: [forms]
      summary: Admin updates a form/poll
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateFormInput"
      responses:
        "200":
          description: Updated form
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FormDetail"
    delete:
      operationId: deleteForm
      tags: [forms]
      summary: Admin deletes a form/poll
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /forms/{id}/vote:
    post:
      operationId: submitVote
      tags: [forms]
      summary: Member votes on a poll
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubmitVoteInput"
      responses:
        "201":
          description: Vote recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FormResponse"
  /forms/{id}/submit:
    post:
      operationId: submitForm
      tags: [forms]
      summary: Member submits a form response
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubmitFormInput"
      responses:
        "201":
          description: Submission recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FormResponse"
  /forms/{id}/responses:
    get:
      operationId: listFormResponses
      tags: [forms]
      summary: Admin views all responses or vote results for a form/poll
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Form responses or poll results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FormResponsesResult"
  /forms/{id}/my-response:
    get:
      operationId: getMyFormResponse
      tags: [forms]
      summary: Check if current user has responded to a form/poll
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User response or null
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MyFormResponse"
components:
  schemas:
...
    FormField:
      type: object
      required: [id, formId, label, fieldType, required, order]
      properties:
        id:
          type: integer
        formId:
          type: integer
        label:
          type: string
        fieldType:
          type: string
          enum: [text, textarea, radio, checkbox, select]
        options:
          type: ["string", "null"]
        required:
          type: boolean
        order:
          type: integer
    PollOption:
      type: object
      required: [id, formId, label, order]
      properties:
        id:
          type: integer
        formId:
          type: integer
        label:
          type: string
        order:
          type: integer
        voteCount:
          type: integer
    FormSummary:
      type: object
      required: [id, title, type, status, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        type:
          type: string
          enum: [form, poll]
        status:
          type: string
          enum: [open, closed]
        createdBy:
          type: integer
        createdByUsername:
          type: ["string", "null"]
        deadline:
          type: ["string", "null"]
        responseCount:
          type: integer
        createdAt:
          type: string
        updatedAt:
          type: string
    FormDetail:
      type: object
      required: [id, title, type, status, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        type:
          type: string
          enum: [form, poll]
        status:
          type: string
          enum: [open, closed]
        createdBy:
          type: integer
        createdByUsername:
          type: ["string", "null"]
        deadline:
          type: ["string", "null"]
        responseCount:
          type: integer
        createdAt:
          type: string
        updatedAt:
          type: string
        fields:
          type: array
          items:
            $ref: "#/components/schemas/FormField"
        options:
          type: array
          items:
            $ref: "#/components/schemas/PollOption"
    CreateFormInput:
      type: object
      required: [title, type]
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 255
        description:
          type: string
        type:
          type: string
          enum: [form, poll]
        deadline:
          type: string
        fields:
          type: array
          items:
            type: object
            required: [label, fieldType]
            properties:
              label:
                type: string
              fieldType:
                type: string
                enum: [text, textarea, radio, checkbox, select]
              options:
                type: string
              required:
                type: boolean
              order:
                type: integer
        options:
          type: array
          items:
            type: object
            required: [label]
            properties:
              label:
                type: string
              order:
                type: integer
    UpdateFormInput:
      type: object
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 255
        description:
          type: string
        status:
          type: string
          enum: [open, closed]
        deadline:
          type: string
    SubmitVoteInput:
      type: object
      required: [optionId]
      properties:
        optionId:
          type: integer
    SubmitFormInput:
      type: object
      required: [answers]
      properties:
        answers:
          type: array
          items:
            type: object
            required: [fieldId, value]
            properties:
              fieldId:
                type: integer
              value:
                type: string
    FormResponse:
      type: object
      required: [id, formId, userId, createdAt]
      properties:
        id:
          type: integer
        formId:
          type: integer
        userId:
          type: integer
        selectedOptionId:
          type: ["integer", "null"]
        createdAt:
          type: string
    FormResponseDetail:
      type: object
      required: [id, userId, username, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        username:
          type: string
        displayName:
          type: ["string", "null"]
        mcUsername:
          type: ["string", "null"]
        selectedOptionId:
          type: ["integer", "null"]
        selectedOptionLabel:
          type: ["string", "null"]
        createdAt:
          type: string
        answers:
          type: array
          items:
            type: object
            required: [fieldId, fieldLabel, value]
            properties:
              fieldId:
                type: integer
              fieldLabel:
                type: string
              value:
                type: string
    FormResponsesResult:
      type: object
      required: [formId, type, responses]
      properties:
        formId:
          type: integer
        type:
          type: string
          enum: [form, poll]
        responses:
          type: array
          items:
            $ref: "#/components/schemas/FormResponseDetail"
        pollResults:
          type: array
          items:
            type: object
            required: [optionId, label, count]
            properties:
              optionId:
                type: integer
              label:
                type: string
              count:
                type: integer
    MyFormResponse:
      type: object
      properties:
        hasResponded:
          type: boolean
        response:
          $ref: "#/components/schemas/FormResponseDetail"
...