# lib/api-zod/src/generated/types/formResponse.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * MC Roleplay Server API
 * OpenAPI spec version: 0.1.0
 */

export interface FormResponse {
  id: number;
  formId: number;
  userId: number;
  /** @nullable */
  selectedOptionId?: number | null;
  createdAt: string;
}