# lib/api-zod/src/generated/types/formSummary.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * MC Roleplay Server API
 * OpenAPI spec version: 0.1.0
 */
import type { FormSummaryStatus } from './formSummaryStatus';
import type { FormSummaryType } from './formSummaryType';

export interface FormSummary {
  id: number;
  title: string;
  /** @nullable */
  description?: string | null;
  type: FormSummaryType;
  status: FormSummaryStatus;
  createdBy?: number;
  /** @nullable */
  createdByUsername?: string | null;
  /** @nullable */
  deadline?: string | null;
  responseCount?: number;
  createdAt: string;
  updatedAt?: string;
}