# lib/api-zod/src/generated/types/pollOption.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * MC Roleplay Server API
 * OpenAPI spec version: 0.1.0
 */

export interface PollOption {
  id: number;
  formId: number;
  label: string;
  order: number;
  voteCount?: number;
}