# lib/api-zod/src/generated/types/submitFormInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * MC Roleplay Server API
 * OpenAPI spec version: 0.1.0
 */
import type { SubmitFormInputAnswersItem } from './submitFormInputAnswersItem';

export interface SubmitFormInput {
  answers: SubmitFormInputAnswersItem[];
}